refactor: 상품 상세 페이지 section 분리
