+++ src/pages/DetailPage/DetailPage.tsx
@@ -1,29 +1,19 @@
-import { FaHeart, FaRegHeart } from "react-icons/fa";
 import { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { useGetProductInfo } from "@/hooks/useGetProductInfo";
 import { useGetProductDetailInfo } from "@/hooks/useGetProductDetailInfo";
 import { useGetProductReviewInfo } from "@/hooks/useGetProductReviewInfo";
 import { useGetProductWishInfo } from "@/hooks/useGetProductWishInfo";
+import ProductInfoSection from "@/pages/DetailPage/components/ProductInfoSection/ProductInfoSection";
+import LikeSection from "@/pages/DetailPage/components/LikeSection/LikeSection";
+import ProductTabSection from "@/pages/DetailPage/components/ProductTabSection/ProductTabSection";
 import NavigationBar from "@/components/NavigationBar/NavigationBar";
 import {
   Wrapper,
   Container,
-  ImageWrapper,
-  Img,
-  Title,
-  Price,
-  BrandWrapper,
-  BrandImg,
-  Brand,
   SectionDivider,
-  TabWrapper,
-  TabButton,
-  Content,
-  LikeSection,
   OrderButton,
   BottomBar,
-  HtmlContentWrapper,
 } from "@/pages/DetailPage/DetailPage.style";
 
 const DetailPage = () => {
@@ -67,148 +57,31 @@ const DetailPage = () => {
     }
   };
 
-  const renderContent = () => {
-    switch (activeTab) {
-      case "description":
-        return (
-          <HtmlContentWrapper
-            dangerouslySetInnerHTML={{ __html: detailInfo?.description ?? "" }}
-          />
-        );
-      case "review":
-        if (isReviewLoading) return <div>리뷰 로딩 중...</div>;
-        if (isReviewError) return <div>리뷰를 불러오지 못했습니다.</div>;
-        if (!reviewData || reviewData.reviews.length === 0)
-          return <div>리뷰가 없습니다.</div>;
-
-        return (
-          <ul style={{ listStyle: "none", padding: 0 }}>
-            {reviewData.reviews.map((review, idx) => {
-              const isLast = idx === reviewData.reviews.length - 1;
-              return (
-                <li
-                  key={review.id}
-                  style={{
-                    padding: "12px",
-                    marginBottom: isLast ? "24px" : "8px",
-                  }}
-                >
-                  <strong
-                    style={{
-                      fontWeight: "bold",
-                      display: "block",
-                      marginBottom: "6px",
-                    }}
-                  >
-                    {review.authorName}
-                  </strong>
-                  <p style={{ whiteSpace: "pre-wrap", margin: 0 }}>
-                    {review.content}
-                  </p>
-                </li>
-              );
-            })}
-          </ul>
-        );
-
-      case "info":
-        if (
-          !detailInfo ||
-          !detailInfo.announcements ||
-          detailInfo.announcements.length === 0
-        )
-          return <div>상세정보가 없습니다.</div>;
-
-        return (
-          <div style={{ padding: "16px", lineHeight: "1.6" }}>
-            {detailInfo.announcements
-              .sort((a, b) => a.displayOrder - b.displayOrder)
-              .map(({ name, value }) => (
-                <div
-                  key={name}
-                  style={{
-                    marginBottom: "16px",
-                  }}
-                >
-                  <div
-                    style={{
-                      fontWeight: "bold",
-                      fontSize: "16px",
-                      color: "#333",
-                      marginBottom: "8px",
-                    }}
-                  >
-                    {name}
-                  </div>
-                  <div
-                    style={{
-                      fontSize: "16px",
-                      color: "#666",
-                      lineHeight: "1.5",
-                    }}
-                  >
-                    {value}
-                  </div>
-                </div>
-              ))}
-          </div>
-        );
-
-      default:
-        return "";
-    }
-  };
-
   return (
     <>
       <NavigationBar />
       <Wrapper>
         <Container>
-          <ImageWrapper>
-            <Img src={item.imageURL} alt={item.name} />
-          </ImageWrapper>
-          <Title>{item.name}</Title>
-          <Price>{item.price.sellingPrice.toLocaleString()}원</Price>
-          <SectionDivider height="1px" />
-          <BrandWrapper>
-            <BrandImg src={item.brandInfo.imageURL} alt={item.brandInfo.name} />
-            <Brand>{item.brandInfo.name}</Brand>
-          </BrandWrapper>
+          <ProductInfoSection item={item} detailInfo={detailInfo} />
         </Container>
         <SectionDivider />
         <Container>
-          <TabWrapper>
-            <TabButton
-              onClick={() => setActiveTab("description")}
-              active={activeTab === "description"}
-            >
-              상품설명
-            </TabButton>
-            <TabButton
-              onClick={() => setActiveTab("review")}
-              active={activeTab === "review"}
-            >
-              선물후기
-            </TabButton>
-            <TabButton
-              onClick={() => setActiveTab("info")}
-              active={activeTab === "info"}
-            >
-              상세정보
-            </TabButton>
-          </TabWrapper>
-          <Content>{renderContent()}</Content>
+          <ProductTabSection
+            activeTab={activeTab}
+            setActiveTab={setActiveTab}
+            detailInfo={detailInfo}
+            reviewData={reviewData}
+            isReviewLoading={isReviewLoading}
+            isReviewError={isReviewError}
+          />
         </Container>
       </Wrapper>
       <BottomBar>
-        <LikeSection onClick={handleLikeToggle}>
-          {liked ? (
-            <FaHeart size={14} color="#f44336" />
-          ) : (
-            <FaRegHeart size={14} color="#999" />
-          )}
-          <div>{likeCount}</div>
-        </LikeSection>
+        <LikeSection
+          isLiked={liked}
+          likeCount={likeCount}
+          onToggle={handleLikeToggle}
+        />
         <OrderButton>주문하기</OrderButton>
       </BottomBar>
     </>
